+++ corehq/apps/case_importer/base.py
@@ -6,6 +6,6 @@ class ImportCases(DataInterface):
     name = ugettext_lazy("Import Cases from Excel")
     slug = "import_cases"
     description = ugettext_lazy("Import case data from an external Excel file")
-    report_template_path = "case_importer_v1/import_cases.html"
+    report_template_path = "case_importer/import_cases.html"
     gide_filters = True
     asynchronous = False
