+++ corehq/apps/case_importer/templates/case_importer/import_cases.html
@@ -3,10 +3,10 @@
 {% load i18n %}
 
 {% block page_content %}
-    {% include 'case_importer_v1/partials/help_message.html' %}
+    {% include 'case_importer/partials/help_message.html' %}
 
     <form class="form-horizontal form-report"
-          action="{% url "corehq.apps.case_importer_v1.views.excel_config" domain %}"
+          action="{% url "corehq.apps.case_importer.views.excel_config" domain %}"
           method="post"
           enctype="multipart/form-data">
         {% csrf_token %}
