+++ settings.py
@@ -260,7 +260,7 @@ HQ_APPS = (
     'corehq.apps.es',
     'corehq.apps.fixtures',
     'corehq.apps.calendar_fixture',
-    'corehq.apps.case_importer_v1',
+    'corehq.apps.case_importer',
     'corehq.apps.reminders',
     'corehq.apps.translations',
     'corehq.apps.tour',
@@ -1237,7 +1237,7 @@ COUCHDB_APPS = [
     'hqcase',
     'hqmedia',
     'hope',
-    'case_importer_v1',
+    'case_importer',
     'indicators',
     'locations',
     'mobile_auth',
