+++ urls.py
@@ -66,7 +66,7 @@ domain_specific = [
     url(r'^case/', include('corehq.apps.case_search.urls')),
     url(r'^cloudcare/', include('corehq.apps.cloudcare.urls')),
     url(r'^fixtures/', include('corehq.apps.fixtures.urls')),
-    url(r'^importer/', include('corehq.apps.case_importer_v1.urls')),
+    url(r'^importer/', include('corehq.apps.case_importer.urls')),
     url(r'^fri/', include('custom.fri.urls')),
     url(r'^ilsgateway/', include('custom.ilsgateway.urls')),
     url(r'^ewsghana/', include('custom.ewsghana.urls')),
